requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.1
supabase==1.0.3
pandas==1.3.4
scikit-learn==1.0.1
matplotlib==3.4.3
numpy==1.21.4